% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genefeat.R
\docType{data}
\name{genefeat}
\alias{genefeat}
\title{Transcript features of the protein coding genes of the human genome}
\format{matrix.}
\source{
\href{http://useast.ensembl.org/index.html}{Ensembl}
}
\usage{
genefeat
}
\description{
Various quantitative features of the transcripts of the set of all 19,962 protein coding genes
in the annotated human genome, according to Ensembl 84.  Features include length of the Coding Sequence (CDS),
lengths of the 5' and 3' UTRs, and the relative usage for each amino acid.
}
\references{
Yates, et. al. (2016) Nucleic Acids Res. 2016 44 Database issue:D710-6. "Ensembl 2016".
(\href{dx.doi.org/10.1093/nar/gkv1157}{doi})
}
\keyword{datasets}
