% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO0007186.R
\docType{data}
\name{GO0007186}
\alias{GO0007186}
\title{The genes pertaining to GO term GO:0007186 (G-protein coupled receptor signaling pathway)}
\format{character vector.}
\source{
\href{http://geneontology.org/}{Gene Ontology}
}
\usage{
GO0007186
}
\description{
The names of all 868 protein coding genes annotated as members of Gene Ontology (GO) term GO:0007186 (G-protein coupled receptor signaling pathway),
according to the Gene Ontology consortium.
}
\references{
Ashburner et al. Gene ontology: tool for the unification of biology (2000) Nat Genet 25(1):25-9.
(\doi{doi:10.1038/75556})
The Gene Ontology Consortium. Gene Ontology Consortium: going forward. (2015) Nucl Acids Res 43 Database issue
	(\href{https://doi.org/10.1093/nar/gku1179}{doi})
}
\keyword{datasets}
