% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpmt.R
\name{logStirling}
\alias{logStirling}
\title{Log Stirling}
\usage{
logStirling(n)
}
\arguments{
\item{n}{integer or vector of integers.}
}
\value{
Stirling approximation of log(n!).  If n <= 14, then computes log(n!) directly, i.e. no Stirling approximation.
}
\description{
Computes the log of the Stirling approximation of n!.
}
