% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpmt.R
\name{minMedian}
\alias{minMedian}
\title{Minimal median}
\usage{
minMedian(x)
}
\arguments{
\item{x}{a vector or matrix of real numbers.  if a matrix, then the minimal median will be computed for each column.}
}
\value{
the minimal median of x.
}
\description{
Computes the minimal median of a vector or matrix.
}
\examples{
minMedian(1:6) # returns 3

}
